Delete the more complex FakeWorkerThreadModule and move it to the appropriate package
