+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/injection/AndroidServicesModuleImplTest.kt
@@ -1,8 +1,8 @@
 package io.embrace.android.embracesdk.injection
 
-import io.embrace.android.embracesdk.FakeWorkerThreadModule
 import io.embrace.android.embracesdk.fakes.injection.FakeCoreModule
 import io.embrace.android.embracesdk.prefs.EmbracePreferencesService
+import io.embrace.android.embracesdk.worker.WorkerThreadModuleImpl
 import org.junit.Assert.assertTrue
 import org.junit.Test
 
@@ -10,10 +10,11 @@ internal class AndroidServicesModuleImplTest {
 
     @Test
     fun testDefault() {
+        val initModule = InitModuleImpl()
         val module = AndroidServicesModuleImpl(
-            initModule = InitModuleImpl(),
+            initModule = initModule,
             coreModule = FakeCoreModule(),
-            workerThreadModule = FakeWorkerThreadModule()
+            workerThreadModule = WorkerThreadModuleImpl(initModule)
         )
 
         assertTrue(module.preferencesService is EmbracePreferencesService)
