+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/injection/CustomerLogModuleImplTest.kt
@@ -1,11 +1,11 @@
 package io.embrace.android.embracesdk.injection
 
-import io.embrace.android.embracesdk.FakeWorkerThreadModule
 import io.embrace.android.embracesdk.fakes.fakeEmbraceSessionProperties
 import io.embrace.android.embracesdk.fakes.injection.FakeAndroidServicesModule
 import io.embrace.android.embracesdk.fakes.injection.FakeCoreModule
 import io.embrace.android.embracesdk.fakes.injection.FakeDeliveryModule
 import io.embrace.android.embracesdk.fakes.injection.FakeEssentialServiceModule
+import io.embrace.android.embracesdk.worker.WorkerThreadModuleImpl
 import org.junit.Assert.assertNotNull
 import org.junit.Test
 
@@ -13,14 +13,15 @@ internal class CustomerLogModuleImplTest {
 
     @Test
     fun testDefaultImplementations() {
+        val initModule = InitModuleImpl()
         val module = CustomerLogModuleImpl(
-            InitModuleImpl(),
+            initModule,
             FakeCoreModule(),
             FakeAndroidServicesModule(),
             FakeEssentialServiceModule(),
             FakeDeliveryModule(),
             fakeEmbraceSessionProperties(),
-            FakeWorkerThreadModule()
+            WorkerThreadModuleImpl(initModule)
         )
 
         assertNotNull(module.networkCaptureService)
