+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/injection/DeliveryModuleImplTest.kt
@@ -1,9 +1,9 @@
 package io.embrace.android.embracesdk.injection
 
-import io.embrace.android.embracesdk.FakeWorkerThreadModule
 import io.embrace.android.embracesdk.fakes.injection.FakeCoreModule
 import io.embrace.android.embracesdk.fakes.injection.FakeEssentialServiceModule
 import io.embrace.android.embracesdk.fakes.injection.FakeStorageModule
+import io.embrace.android.embracesdk.worker.WorkerThreadModuleImpl
 import org.junit.Assert.assertNotNull
 import org.junit.Test
 
@@ -13,7 +13,7 @@ internal class DeliveryModuleImplTest {
     fun testDefaultImplementations() {
         val module = DeliveryModuleImpl(
             FakeCoreModule(),
-            FakeWorkerThreadModule(),
+            WorkerThreadModuleImpl(InitModuleImpl()),
             FakeStorageModule(),
             FakeEssentialServiceModule(),
         )
