+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/internal/crash/LastRunCrashVerifierTest.kt
@@ -1,6 +1,7 @@
 package io.embrace.android.embracesdk.internal.crash
 
-import io.embrace.android.embracesdk.FakeWorkerThreadModule
+import io.embrace.android.embracesdk.fakes.injection.FakeInitModule
+import io.embrace.android.embracesdk.fakes.injection.FakeWorkerThreadModule
 import io.embrace.android.embracesdk.worker.BackgroundWorker
 import io.embrace.android.embracesdk.worker.WorkerName
 import io.mockk.every
@@ -23,7 +24,7 @@ internal class LastRunCrashVerifierTest {
     fun setUp() {
         mockCrashFileMarker = mockk()
         lastRunCrashVerifier = LastRunCrashVerifier(mockCrashFileMarker)
-        fakeWorkerThreadModule = FakeWorkerThreadModule()
+        fakeWorkerThreadModule = FakeWorkerThreadModule(fakeInitModule = FakeInitModule(), name = WorkerName.BACKGROUND_REGISTRATION)
         worker = fakeWorkerThreadModule.backgroundWorker(WorkerName.BACKGROUND_REGISTRATION)
     }
