+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/ndk/NativeModuleImplTest.kt
@@ -1,12 +1,13 @@
 package io.embrace.android.embracesdk.ndk
 
-import io.embrace.android.embracesdk.FakeWorkerThreadModule
 import io.embrace.android.embracesdk.fakes.fakeEmbraceSessionProperties
 import io.embrace.android.embracesdk.fakes.injection.FakeAndroidServicesModule
 import io.embrace.android.embracesdk.fakes.injection.FakeCoreModule
 import io.embrace.android.embracesdk.fakes.injection.FakeDeliveryModule
 import io.embrace.android.embracesdk.fakes.injection.FakeEssentialServiceModule
 import io.embrace.android.embracesdk.fakes.injection.FakeStorageModule
+import io.embrace.android.embracesdk.injection.InitModuleImpl
+import io.embrace.android.embracesdk.worker.WorkerThreadModuleImpl
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Test
@@ -22,7 +23,7 @@ internal class NativeModuleImplTest {
             FakeDeliveryModule(),
             FakeAndroidServicesModule(),
             fakeEmbraceSessionProperties(),
-            FakeWorkerThreadModule()
+            WorkerThreadModuleImpl(InitModuleImpl())
         )
         assertNotNull(module.ndkService)
         assertNull(module.nativeThreadSamplerService)
